+++ driver/web/apis_system.go
@@ -790,37 +790,51 @@ func (h SystemApisHandler) createApplication(l *logs.Log, r *http.Request, claim
 		return l.HttpResponseErrorAction(logutils.ActionRead, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, false)
 	}
 
-	var requestData Def.SystemReqCreateApplication
+	var requestData Def.Application
 	err = json.Unmarshal(data, &requestData)
 	if err != nil {
 		return l.HttpResponseErrorAction(logutils.ActionUnmarshal, model.TypeApplication, nil, err, http.StatusBadRequest, true)
 	}
 
-	name := requestData.Name
-	multiTenant := requestData.MultiTenant
-	admin := requestData.Admin
-	sharedIdentities := requestData.SharedIdentities
-	applicationTypes := requestData.ApplicationTypes
-
 	appTypes := make([]model.ApplicationType, 0)
-	if applicationTypes != nil {
-		for _, at := range *applicationTypes {
-			versions := make([]model.Version, 0)
-			if at.Versions != nil {
-				for _, v := range *at.Versions {
-					versionNumbers := model.VersionNumbersFromString(v)
-					if versionNumbers != nil {
-						versions = append(versions, model.Version{VersionNumbers: *versionNumbers})
-					}
-				}
-			}
-			appTypes = append(appTypes, model.ApplicationType{Identifier: at.Identifier, Name: *at.Name, Versions: versions})
-		}
+	if requestData.Types != nil {
+		appTypes = applicationTypeListFromDef(*requestData.Types)
 	}
 
-	_, err = h.coreAPIs.System.SysCreateApplication(name, multiTenant, admin, sharedIdentities, appTypes)
+	_, err = h.coreAPIs.System.SysCreateApplication(requestData.Name, requestData.MultiTenant, requestData.Admin, requestData.SharedIdentities, appTypes)
 	if err != nil {
 		return l.HttpResponseErrorAction(logutils.ActionCreate, model.TypeApplication, nil, err, http.StatusInternalServerError, true)
+	}
+
+	return l.HttpResponseSuccess()
+}
+
+func (h SystemApisHandler) updateApplication(l *logs.Log, r *http.Request, claims *tokenauth.Claims) logs.HttpResponse {
+	params := mux.Vars(r)
+	ID := params["id"]
+	if len(ID) <= 0 {
+		return l.HttpResponseErrorData(logutils.StatusMissing, logutils.TypeQueryParam, logutils.StringArgs("id"), nil, http.StatusBadRequest, false)
+	}
+
+	data, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionRead, logutils.TypeRequestBody, nil, err, http.StatusBadRequest, false)
+	}
+
+	var requestData Def.Application
+	err = json.Unmarshal(data, &requestData)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionUnmarshal, model.TypeApplication, nil, err, http.StatusBadRequest, true)
+	}
+
+	appTypes := make([]model.ApplicationType, 0)
+	if requestData.Types != nil {
+		appTypes = applicationTypeListFromDef(*requestData.Types)
+	}
+
+	err = h.coreAPIs.System.SysUpdateApplication(ID, requestData.Name, requestData.MultiTenant, requestData.Admin, requestData.SharedIdentities, appTypes)
+	if err != nil {
+		return l.HttpResponseErrorAction(logutils.ActionUpdate, model.TypeApplication, nil, err, http.StatusInternalServerError, true)
 	}
 
 	return l.HttpResponseSuccess()
@@ -831,11 +845,7 @@ func (h SystemApisHandler) getApplications(l *logs.Log, r *http.Request, claims 
 	if err != nil {
 		return l.HttpResponseErrorAction(logutils.ActionGet, model.TypeApplication, nil, err, http.StatusInternalServerError, true)
 	}
-	var response []Def.ApplicationFields
-	for _, application := range applications {
-		r := applicationToDef(application)
-		response = append(response, r)
-	}
+	response := applicationsToDef(applications)
 
 	data, err := json.Marshal(response)
 	if err != nil {