+++ CHANGELOG.md
@@ -6,11 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
-## [1.26.0] - 2022-10-05
 ### Added
 - Add scopes to service accounts [#605](https://github.com/rokwire/core-building-block/issues/605)
+- Update application system API [#549](https://github.com/rokwire/core-building-block/issues/549)
+- Update group admin API [#518](https://github.com/rokwire/core-building-block/issues/518)
 
+## [1.26.0] - 2022-10-05
 ### Fixed
 - Services crashes on anonymous login [#603](https://github.com/rokwire/core-building-block/issues/603)